Add `graphql` export to `.keystone/types`
