+++ .changeset/bright-pots-pretend.md
@@ -1,0 +1,5 @@
+---
+'@keystone-6/core': minor
+---
+
+Adds `graphql` export to `.keystone/types` that is the same as the `graphql` export from `@keystone-6/core` but uses the specific generated context type for the project
