+++ examples/extend-graphql-schema-graphql-ts/custom-schema.ts
@@ -1,5 +1,4 @@
-import { graphql } from '@keystone-6/core';
-import { Context } from '.keystone/types';
+import { graphql, Context } from '.keystone/types';
 
 export const extendGraphqlSchema = graphql.extend(base => {
   const Statistics = graphql.object<{ authorId: string }>()({
