+++ packages/core/___internal-do-not-use-will-break-in-patch/graphql-ts/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-graphql-ts.cjs.js",
+  "module": "dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-graphql-ts.esm.js"
+}
