+++ packages/core/___internal-do-not-use-will-break-in-patch/load-config/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-load-config.cjs.js",
+  "module": "dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-load-config.esm.js"
+}
