--- packages/core/___internal-do-not-use-will-break-in-patch/require-source/package.json
@@ -1,4 +1,0 @@
-{
-  "main": "dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-require-source.cjs.js",
-  "module": "dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-require-source.esm.js"
-}
