+++ packages/core/src/___internal-do-not-use-will-break-in-patch/graphql-ts-without-context.ts
@@ -1,0 +1,1 @@
+export * from '../types/schema/api-without-context';
