+++ packages/core/src/___internal-do-not-use-will-break-in-patch/load-config.ts
@@ -1,0 +1,1 @@
+export { loadConfig } from '../lib/config/loadConfig';
