--- packages/core/src/___internal-do-not-use-will-break-in-patch/require-source.ts
@@ -1,1 +1,0 @@
-export { requireSource } from '../lib/config/requireSource';
