+++ packages/core/src/artifacts.ts
@@ -203,6 +203,57 @@ const nextGraphQLAPIDTS = `export const config: any;
 export default config;
 `;
 
+const graphqlTsAPI = `import * as graphqlTsSchema from '@keystone-6/core/___internal-do-not-use-will-break-in-patch/graphql-ts';
+import type { Context } from '../types';
+export * from '@keystone-6/core/___internal-do-not-use-will-break-in-patch/graphql-ts-without-context';
+export { field, fields, interface, interfaceField, object, union } from './graphql-ts-with-context';
+
+export type { Context }
+
+export type NullableType = graphqlTsSchema.NullableType<Context>;
+export type Type = graphqlTsSchema.Type<Context>;
+export type NullableOutputType = graphqlTsSchema.NullableOutputType<Context>;
+export type OutputType = graphqlTsSchema.OutputType<Context>;
+export type Field<
+  Source,
+  Args extends Record<string, graphqlTsSchema.Arg<any>>,
+  TType extends OutputType,
+  Key extends string
+> = graphqlTsSchema.Field<Source, Args, TType, Key, Context>;
+export type FieldResolver<
+  Source,
+  Args extends Record<string, graphqlTsSchema.Arg<any>>,
+  TType extends OutputType
+> = graphqlTsSchema.FieldResolver<Source, Args, TType, Context>;
+export type ObjectType<Source> = graphqlTsSchema.ObjectType<Source, Context>;
+export type UnionType<Source> = graphqlTsSchema.UnionType<Source, Context>;
+export type InterfaceType<
+  Source,
+  Fields extends Record<string, graphqlTsSchema.InterfaceField<any, OutputType, Context>>
+> = graphqlTsSchema.InterfaceType<Source, Fields, Context>;
+export type InterfaceField<
+  Args extends Record<string, graphqlTsSchema.Arg<any>>,
+  TType extends OutputType
+> = graphqlTsSchema.InterfaceField<Args, TType, Context>;
+`;
+
+// this whole export = thing is just so that ts retains JSDoc comments (the alternative of creating a variable does not preserve them)
+const graphqlTsWithContextAPI = `import { GraphQLSchemaAPIWithContext } from '@keystone-6/core/___internal-do-not-use-will-break-in-patch/graphql-ts';
+import { Context } from './graphql-ts';
+
+declare const __graphql: GraphQLSchemaAPIWithContext<Context>;
+
+export = __graphql;
+`;
+
+export async function generateTypesJSFile(cwd: string) {
+  const dotKeystoneDir = path.join(cwd, 'node_modules/.keystone');
+  await fs.outputFile(
+    path.join(dotKeystoneDir, 'types.js'),
+    "exports.graphql = require('@keystone-6/core').graphql;"
+  );
+}
+
 export async function generateNodeModulesArtifactsWithoutPrismaClient(
   graphQLSchema: GraphQLSchema,
   config: KeystoneConfig,
@@ -216,7 +267,12 @@ export async function generateNodeModulesArtifactsWithoutPrismaClient(
       path.join(dotKeystoneDir, 'types.d.ts'),
       printGeneratedTypes(graphQLSchema, lists)
     ),
-    fs.outputFile(path.join(dotKeystoneDir, 'types.js'), ''),
+    fs.outputFile(path.join(dotKeystoneDir, 'internal/graphql-ts.d.ts'), graphqlTsAPI),
+    fs.outputFile(
+      path.join(dotKeystoneDir, 'internal/graphql-ts-with-context.d.ts'),
+      graphqlTsWithContextAPI
+    ),
+    generateTypesJSFile(cwd),
     ...(config.experimental?.generateNodeAPI
       ? [
           fs.outputFile(path.join(dotKeystoneDir, 'api.js'), nodeAPIJS(cwd, config)),
