+++ packages/core/src/lib/config/loadConfig.ts
@@ -1,0 +1,10 @@
+import path from 'path';
+import { generateTypesJSFile } from '../../artifacts';
+import { KeystoneConfig } from '../../types';
+import { initConfig } from './initConfig';
+import { requireSource } from './requireSource';
+
+export async function loadConfig(cwd: string): Promise<KeystoneConfig> {
+  await generateTypesJSFile(cwd);
+  return initConfig(requireSource(path.join(cwd, 'keystone')).default);
+}
