+++ packages/core/src/scripts/run/dev.ts
@@ -5,11 +5,11 @@ import express from 'express';
 import { GraphQLSchema, printSchema } from 'graphql';
 import fs from 'fs-extra';
 import chalk from 'chalk';
+import { loadConfig } from '../../lib/config/loadConfig';
 import { generateAdminUI } from '../../admin-ui/system';
 import { devMigrations, pushPrismaSchemaToDatabase } from '../../lib/migrations';
 import { createSystem } from '../../lib/createSystem';
 import { initConfig } from '../../lib/config/initConfig';
-import { requireSource } from '../../lib/config/requireSource';
 import { defaults } from '../../lib/config/defaults';
 import { createExpressServer } from '../../lib/server/createExpressServer';
 import { createAdminUIMiddleware } from '../../lib/server/createAdminUIMiddleware';
@@ -21,7 +21,7 @@ import {
   getSchemaPaths,
   requirePrismaClient,
 } from '../../artifacts';
-import { getAdminPath, getConfigPath } from '../utils';
+import { getAdminPath } from '../utils';
 import { AdminMetaRootVal, CreateContext, KeystoneConfig } from '../../types';
 import { serializePathForImport } from '../../admin-ui/utils/serializePathForImport';
 import { initialiseLists } from '../../lib/core/types-for-lists';
@@ -50,7 +50,7 @@ export const dev = async (cwd: string, shouldDropDatabase: boolean) => {
   // - you have an error in your config after startup -> will keep the last working version until importing the config succeeds
   // also, if you're thinking "why not always use the Next api route to get the config"?
   // this will get the GraphQL API up earlier
-  const config = initConfig(requireSource(getConfigPath(cwd)).default);
+  const config = await loadConfig(cwd);
 
   const isReady = () =>
     expressServer !== null && (hasAddedAdminUIMiddleware || config.ui?.isDisabled === true);
